pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["gabrielsrochaa <gabriel.srocha97@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.4"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
python-decouple = "^3.8"
pytest = "^7.3.0"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
